Refactor two-factor authentication service and make 2FA login logic in the RESTful API
