+++ package.json
@@ -50,6 +50,8 @@
     "@types/jest": "^29.5.2",
     "@types/multer": "^1.4.7",
     "@types/node": "^20.5.6",
+    "@types/passport-jwt": "^3.0.9",
+    "@types/qrcode": "^1.5.1",
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
