+++ src/auth/auth.module.ts
@@ -9,6 +9,7 @@ import { Intra42Strategy } from './strategies/intra42.strategy';
 import { AuthController } from './auth.controller';
 import { UserService } from 'src/user/user.service';
 import { UserRelationsService } from 'src/user/services/user-relations.service';
+import { Auth2faService } from './services/auth-2fa.service';
 
 /**
  * Module for handling authentication-related functionality.
@@ -24,11 +25,13 @@ import { UserRelationsService } from 'src/user/services/user-relations.service';
   providers: [
     AuthResolver, // GraphQL resolver for authentication-related functionality
     AuthService, // Service for handling authentication-related functionality
+    Auth2faService, // Service for handling 2FA-related functionality
     JwtService, // Service for handling JWTs
     PrismaService, // Service for handling database interactions
     UserService, // Service for handling user-related functionality
     UserRelationsService, // Service for handling user relations (follows, block, etc.)
     Intra42Strategy, // Passport strategy for authenticating with 42 intra
   ],
+  exports: [Auth2faService], // Export Auth2faService to be used in other modules
 })
 export class AuthModule {}
