+++ src/auth/auth.resolver.ts
@@ -1,12 +1,8 @@
 import { Resolver, Mutation, Args } from '@nestjs/graphql';
-import { ForbiddenException, Res } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { Response } from 'express';
+import { ForbiddenException } from '@nestjs/common';
 
 import { Auth } from './entities/auth.entity';
-import { AuthService } from './auth.service';
-import { SignIn2faInput } from './dto/signin-2fa.input';
-import { SignResponse } from './dto/sign.response';
+import { Auth2faService } from './services/auth-2fa.service';
 
 /**
  * Resolver class for handling GraphQL queries and mutations related to authentication.
@@ -19,52 +15,17 @@ export class AuthResolver {
   /**
    * Creates an instance of the AuthResolver class.
    *
-   * @param {AuthService} authService - The authentication service used for resolving authentication-related queries and mutations.
+   * @param {Auth2faService} auth2faService - The service responsible for two-factor authentication (2FA) related functionality.
    */
-  constructor(
-    private readonly authService: AuthService,
-    private readonly configService: ConfigService,
-  ) {}
+  constructor(private readonly auth2faService: Auth2faService) {}
 
-  @Mutation(() => Auth, {
-    name: 'signin2fa',
-    description: 'Signs in a user with 2FA',
-  })
-  async signin2fa(
-    @Args('signIn2faInput', { type: () => SignIn2faInput })
-    signInInput: SignIn2faInput,
-    @Res() res: Response,
-  ): Promise<void> {
-    /**
-     * Verify the 2FA token.
-     * If the token is invalid, throw an error.
-     * Otherwise, sign in the user.
-     */
-    const isTokenValid: boolean = await this.authService.verify2faToken(
-      signInInput.userId,
-      signInInput.token,
-    );
-
-    if (!isTokenValid)
-      throw new ForbiddenException('Wrong authentication token');
-
-    const response: SignResponse =
-      await this.authService.signin2fa(signInInput);
-
-    // Set the access token as a cookie
-    res.cookie('accessToken', response.accessToken, {
-      // Cookie is not accessible via client-side JavaScript for better security
-      httpOnly: true,
-      // Cookie is only sent to the server via HTTPS
-      secure: true,
-    });
-
-    console.log(response.accessToken);
-
-    // Redirect the user to the frontend
-    res.redirect(this.configService.get<string>('FRONTEND_URL'));
-  }
-
+  /**
+   * Mutation for generating a 2FA token for the current user.
+   *
+   * @param {string} userId - User ID.
+   * @param {string} username - Username of the user.
+   * @returns {Promise<Auth>} - The generated 2FA token and OTP authentication URL.
+   */
   @Mutation(() => Auth, {
     name: 'generate2faToken',
     description: 'Generates a 2FA token for the current user',
@@ -73,16 +34,24 @@ export class AuthResolver {
     @Args('userId', { type: () => String }) userId: string,
     @Args('username', { type: () => String }) username: string,
   ): Promise<Auth> {
-    const { otpAuthUrl } = await this.authService.generate2faSecret(
+    const { otpAuthUrl } = await this.auth2faService.generate2faSecret(
       userId,
       username,
     );
     return {
       status: '2FA token generated',
-      otpAuthUrl: await this.authService.generateQrCodeDataURL(otpAuthUrl),
+      otpAuthUrl: await this.auth2faService.generateQrCodeDataURL(otpAuthUrl),
     };
   }
 
+  /**
+   * Mutation for turning on two-factor authentication (2FA) for the current user.
+   *
+   * @param {string} userId - User ID.
+   * @param {string} token - Authentication token for verifying 2FA setup.
+   * @returns {Promise<Auth>} - The status and updated user information.
+   * @throws {ForbiddenException} - If the provided token is invalid.
+   */
   @Mutation(() => Auth, {
     name: 'turnOn2fa',
     description: 'Turns on 2FA for the current user',
@@ -91,7 +60,7 @@ export class AuthResolver {
     @Args('userId', { type: () => String }) userId: string,
     @Args('token', { type: () => String }) token: string,
   ): Promise<Auth> {
-    const isTokenValid: boolean = await this.authService.verify2faToken(
+    const isTokenValid: boolean = await this.auth2faService.verify2faToken(
       userId,
       token,
     );
@@ -101,10 +70,18 @@ export class AuthResolver {
 
     return {
       status: '2FA turned on',
-      user: await this.authService.turnOn2fa(userId),
+      user: await this.auth2faService.turnOn2fa(userId),
     };
   }
 
+  /**
+   * Mutation for turning off two-factor authentication (2FA) for the current user.
+   *
+   * @param {string} userId - User ID.
+   * @param {string} token - Authentication token for verifying 2FA removal.
+   * @returns {Promise<Auth>} - The status and updated user information.
+   * @throws {ForbiddenException} - If the provided token is invalid.
+   */
   @Mutation(() => Auth, {
     name: 'turnOff2fa',
     description: 'Turns off 2FA for the current user',
@@ -113,7 +90,7 @@ export class AuthResolver {
     @Args('userId', { type: () => String }) userId: string,
     @Args('token', { type: () => String }) token: string,
   ): Promise<Auth> {
-    const isTokenValid: boolean = await this.authService.verify2faToken(
+    const isTokenValid: boolean = await this.auth2faService.verify2faToken(
       userId,
       token,
     );
@@ -123,10 +100,17 @@ export class AuthResolver {
 
     return {
       status: '2FA turned off',
-      user: await this.authService.turnOff2fa(userId),
+      user: await this.auth2faService.turnOff2fa(userId),
     };
   }
 
+  /**
+   * Mutation for verifying a two-factor authentication (2FA) token for the current user.
+   *
+   * @param {string} userId - User ID.
+   * @param {string} token - 2FA token to verify.
+   * @returns {Promise<Auth>} - The status and whether the 2FA token is valid.
+   */
   @Mutation(() => Auth, {
     name: 'verify2fa',
     description: 'Verifies a 2FA token for the current user',
@@ -137,7 +121,7 @@ export class AuthResolver {
   ): Promise<Auth> {
     return {
       status: '2FA token verified',
-      is2faValid: await this.authService.verify2faToken(userId, token),
+      is2faValid: await this.auth2faService.verify2faToken(userId, token),
     };
   }
 }
